requirements.txt
# Core dependencies
python-dotenv==1.0.0
pillow==10.2.0
numpy==1.26.3

# LLM API clients
google-generativeai==0.3.2
anthropic==0.18.1
openai==1.12.0

# Data handling
pandas==2.2.0
openpyxl==3.1.2

# Analysis and visualization
matplotlib==3.8.2
seaborn==0.13.1
scipy==1.12.0

# Jupyter
jupyter==1.0.0
ipykernel==6.29.0

# Utilities
tqdm==4.66.1
requests==2.31.0
